requirements.txt
numpy==2.2.0
opencv-python==4.10.0.84
pillow==11.0.0
